write local plugin in typescript
