+++ examples/fishjam-chat/tsconfig.json
@@ -2,6 +2,5 @@
   "extends": "expo/tsconfig.base",
   "compilerOptions": {
     "strict": true
-  },
-  "include": ["./plugin"]
+  }
 }
